renamed partition_and_pq to partition
