+++ src/CMakeLists.txt
@@ -9,8 +9,8 @@ else()
     #file(GLOB CPP_SOURCES *.cpp)
     set(CPP_SOURCES ann_exception.cpp aux_utils.cpp distance.cpp index.cpp
         linux_aligned_file_reader.cpp math_utils.cpp natural_number_map.cpp
-        natural_number_set.cpp memory_mapper.cpp partition_and_pq.cpp
-        pq_flash_index.cpp pq.cpp logger.cpp utils.cpp)
+        natural_number_set.cpp memory_mapper.cpp partition.cpp pq.cpp
+        pq_flash_index.cpp logger.cpp utils.cpp)
     add_library(${PROJECT_NAME} ${CPP_SOURCES})
     add_library(${PROJECT_NAME}_s STATIC ${CPP_SOURCES})
 endif()
