+++ tests/utils/partition_data.cpp
@@ -4,7 +4,7 @@
 #include <index.h>
 #include <math_utils.h>
 #include "cached_io.h"
-#include "partition_and_pq.h"
+#include "partition.h"
 
 // DEPRECATED: NEED TO REPROGRAM
